// backend/shared/type.ts
export type hotelType = {
    userId: string;
    name: string;
    city: string;
    country: string;
    description: string;
    pricePerNight: number;
    starRating: number;
    type: string;
    facilities: Array<string>;
    adultCount: number;
    childCount: number;
    images: Array<string>;
};